+++ internal/acceptance/openstack/networking/v2/conditions.go
@@ -2,6 +2,7 @@ package v2
 
 import (
 	"context"
+	"os"
 	"testing"
 
 	"github.com/gophercloud/gophercloud/v2"
@@ -14,5 +15,13 @@ func RequireNeutronExtension(t *testing.T, client *gophercloud.ServiceClient, ex
 	_, err := extensions.Get(context.TODO(), client, extension).Extract()
 	if err != nil {
 		t.Skipf("this test requires %s Neutron extension", extension)
+	}
+}
+
+// RequirePortForwarding will restrict a test to only be run in environments
+// that support port forwarding
+func RequirePortForwarding(t *testing.T) {
+	if os.Getenv("OS_PORTFORWARDING_ENVIRONMENT") == "" {
+		t.Skip("this test requires support for port forwarding")
 	}
 }
