+++ internal/acceptance/openstack/networking/v2/extensions/layer3/portforwardings_test.go
@@ -15,7 +15,7 @@ import (
 )
 
 func TestLayer3PortForwardingsCreateDelete(t *testing.T) {
-	clients.RequirePortForwarding(t)
+	networking.RequirePortForwarding(t)
 
 	client, err := clients.NewNetworkV2Client()
 	th.AssertNoErr(t, err)
